Extract HeroCanvas component
